Move visualise to orchestrator
